chore: rename `open-in-brower.ts` to `open-in-browser.ts`
^ Like it says.
