+++ .changeset/grumpy-walls-nail.md
@@ -1,0 +1,5 @@
+---
+"wrangler": patch
+---
+
+chore: rename `open-in-brower.ts` to `open-in-browser.ts`
